remove UserManager.php
