+++ api/src/Controllers/UserStatusController.php
@@ -14,7 +14,7 @@ class UserStatusController
         ]);
 
         /** @var \Matcha\Api\Model\User */
-        $user = Flight::user()->model();
+        $user = Flight::user();
 
         $user->online = Flight::request()->data->state;
         $user->save();
